realty_price_mini/requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.28
kagglehub>=0.2
pyyaml>=6.0
matplotlib>=3.7
seaborn>=0.13
scikit-learn>=1.3
